Create check_open_status_&_delete_wb.bas
In this example, the Workbooks.Open method is used to try to open the workbook "C:\MyFolder\MyWorkbook.xlsx". If the workbook is already open, the Workbooks.Open method will return a reference to the open workbook. If the workbook is not open, the Workbooks.Open method will return Nothing.

The On Error Resume Next statement is used to suppress any error messages that may be generated when the Workbooks.Open method is called. The On Error GoTo 0 statement is used to turn off error handling.

The If statement is then used to check if the wb variable is Nothing. If wb is Nothing, it means that the workbook is closed, and the Kill function is used to delete the workbook. If wb is not Nothing, it means that the workbook is open, and the Close method of the Workbook object is used to close the workbook before deleting it using the Kill function.

Note that this VBA code will only work if you are running it from a Windows machine. If you are running it from another operating system, you may need to modify the path to the workbook to use a path that is valid on that operating system.