Create check_open_status_&_delete_wb.vbs
In this VBScript code, the CreateObject function is used to create an Excel application object, which allows you to access the Excel object model. The Workbooks.Open method of the Excel application object is then used to try to open the workbook "C:\MyFolder\MyWorkbook.xlsx".

The On Error Resume Next statement is used to suppress any error messages that may be generated when the Workbooks.Open method is called. The On Error GoTo 0 statement is used to turn off error handling.

The If statement is then used to check if the wb variable is Nothing. If wb is Nothing, it means that the workbook is closed, and the Kill function is used to delete the workbook. If wb is not Nothing, it means that the workbook is open, and the Close method of the Workbook object is used to close the workbook before deleting it using the Kill function.

Finally, the Set statements are used to set the wb and xlApp variables to Nothing to release the memory used by these objects.

Note that this VBScript code will only work if you are running it from a Windows machine. If you are running it from another operating system, you may need to modify the path to the workbook to use a path that is valid on that operating system.